+++ react-server-dom-vite-example/src/entry.client.tsx
@@ -1,7 +1,7 @@
 import ReactClient from "@jacob-ebey/react-server-dom-vite/client";
 import ReactDomClient from "react-dom/client";
 import type { ServerPayload } from "./entry.rsc";
-import { resolveClientReference } from "./utils/client";
+import { resolveClientReference } from "./utils/client-reference";
 import { getFlightStreamBrowser } from "./utils/stream-script";
 
 async function main() {
