+++ react-server-dom-vite-example/src/utils/client-reference.ts
