+++ react-server-dom-vite-example/src/utils/fetch.ts
